some cosmetics changes + energyRecoil+ digits format

+++ src/main/java/ru/mauveferret/ParticleInMatterCalculator.java
@@ -16,7 +16,7 @@ public abstract class ParticleInMatterCalculator {
 
     //how mich lines per FileRead can be processed
     public int stringCountPerCycle = 2000;
-    long time;
+    double time;
 
 
     //primary beam
